Rename `useLogs` to `use-logs`
